+++ src/Datasets.php
@@ -74,24 +74,45 @@ final class Datasets
             $data = iterator_to_array($data);
         }
 
-        $namedData = [];
-        foreach ($data as $values) {
+        $dataSetDescriptions = [];
+        $dataSetValues       = [];
+
+        foreach ($data as $key => $values) {
             $values = is_array($values) ? $values : [$values];
 
-            $name             = $description . self::getDataSetDescription($values);
-            $namedData[$name] = $values;
+            $dataSetDescriptions[] = $description . self::getDataSetDescription($key, $values);
+            $dataSetValues[]       = $values;
+        }
+
+        foreach (array_count_values($dataSetDescriptions) as $descriptionToCheck => $count) {
+            if ($count > 1) {
+                $index = 1;
+                foreach ($dataSetDescriptions as $i => $dataSetDescription) {
+                    if ($dataSetDescription === $descriptionToCheck) {
+                        $dataSetDescriptions[$i] .= sprintf(' #%d', $index++);
+                    }
+                }
+            }
+        }
+
+        $namedData = [];
+        foreach ($dataSetDescriptions as $i => $dataSetDescription) {
+            $namedData[$dataSetDescription] = $dataSetValues[$i];
         }
 
         return $namedData;
     }
 
     /**
+     * @param int|string        $key
      * @param array<int, mixed> $data
      */
-    private static function getDataSetDescription(array $data): string
+    private static function getDataSetDescription($key, array $data): string
     {
         $exporter = new Exporter();
 
-        return \sprintf(' with (%s)', $exporter->shortenedRecursiveExport($data));
+        $nameInsert = is_string($key) ? \sprintf('data set "%s" ', $key) : '';
+
+        return \sprintf(' with %s(%s)', $nameInsert, $exporter->shortenedRecursiveExport($data));
     }
 }