+++ src/Support/Reflection.php
@@ -6,9 +6,12 @@ namespace Pest\Support;
 
 use Closure;
 use Pest\Exceptions\ShouldNotHappen;
+use Pest\TestSuite;
 use ReflectionClass;
 use ReflectionException;
 use ReflectionFunction;
+use ReflectionNamedType;
+use ReflectionParameter;
 use ReflectionProperty;
 
 /**
@@ -27,11 +30,25 @@ final class Reflection
     {
         $reflectionClass = new ReflectionClass($object);
 
-        $reflectionMethod = $reflectionClass->getMethod($method);
+        try {
+            $reflectionMethod = $reflectionClass->getMethod($method);
 
-        $reflectionMethod->setAccessible(true);
+            $reflectionMethod->setAccessible(true);
 
-        return $reflectionMethod->invoke($object, ...$args);
+            return $reflectionMethod->invoke($object, ...$args);
+        } catch (ReflectionException $exception) {
+            if (method_exists($object, '__call')) {
+                return $object->__call($method, $args);
+            }
+
+            if (is_callable($method)) {
+                return Closure::fromCallable($method)->bindTo(
+                    TestSuite::getInstance()->test
+                )(...$args);
+            }
+
+            throw $exception;
+        }
     }
 
     /**
@@ -108,5 +125,33 @@ final class Reflection
 
         $reflectionProperty->setAccessible(true);
         $reflectionProperty->setValue($object, $value);
+    }
+
+    /**
+     * Get the class name of the given parameter's type, if possible.
+     *
+     * @see https://github.com/laravel/framework/blob/v6.18.25/src/Illuminate/Support/Reflector.php
+     */
+    public static function getParameterClassName(ReflectionParameter $parameter): ?string
+    {
+        $type = $parameter->getType();
+
+        if (!$type instanceof ReflectionNamedType || $type->isBuiltin()) {
+            return null;
+        }
+
+        $name = $type->getName();
+
+        if (($class = $parameter->getDeclaringClass()) instanceof ReflectionClass) {
+            if ($name === 'self') {
+                return $class->getName();
+            }
+
+            if ($name === 'parent' && ($parent = $class->getParentClass()) instanceof ReflectionClass) {
+                return $parent->getName();
+            }
+        }
+
+        return $name;
     }
 }