+++ tests/Features/Helpers.php
@@ -7,7 +7,7 @@ function addUser()
 
 it('can set/get properties on $this', function () {
     addUser();
-    assertEquals('nuno', $this->user);
+    expect($this->user)->toBe('nuno');
 });
 
 it('throws error if property do not exist', function () {
@@ -27,17 +27,18 @@ function mockUser()
     $mock = test()->createMock(User::class);
 
     $mock->method('getName')
-         ->willReturn('maduro');
+        ->willReturn('maduro');
 
     return $mock;
 }
 
 it('allows to call underlying protected/private methods', function () {
     $user = mockUser();
-
-    assertEquals('maduro', $user->getName());
+    expect($user->getName())->toBe('maduro');
 });
 
 it('throws error if method do not exist', function () {
     test()->name();
 })->throws(\ReflectionException::class, 'Call to undefined method PHPUnit\Framework\TestCase::name()');
+
+it('can forward unexpected calls to any global function')->_assertThat();