+++ tests/Features/HigherOrderMessages.php
@@ -1,7 +1,0 @@
-<?php
-
-beforeEach()->assertTrue(true);
-
-it('proxies calls to object')->assertTrue(true);
-
-afterEach()->assertTrue(true);