+++ tests/Features/HigherOrderTests.php
@@ -1,0 +1,11 @@
+<?php
+
+beforeEach()->assertTrue(true);
+
+it('proxies calls to object')->assertTrue(true);
+
+it('is capable doing multiple assertions')
+    ->assertTrue(true)
+    ->assertFalse(false);
+
+afterEach()->assertTrue(true);