+++ tests/Visual/SingleTestOrDirectory.php
@@ -2,31 +2,44 @@
 
 use Symfony\Component\Process\Process;
 
-$run = function (string $target) {
-    $process = new Process(['./bin/pest', $target], dirname(__DIR__, 2));
+$run = function (string $target, $decorated = false) {
+    $process = new Process(['php', 'bin/pest', $target], dirname(__DIR__, 2));
 
     $process->run();
 
-    return preg_replace('#\\x1b[[][^A-Za-z]*[A-Za-z]#', '', $process->getOutput());
+    return $decorated ? $process->getOutput() : preg_replace('#\\x1b[[][^A-Za-z]*[A-Za-z]#', '', $process->getOutput());
 };
 
-test('allows to run a single test', function () use ($run) {
-    assertStringContainsString(<<<EOF
-   PASS  Tests\Fixtures\DirectoryWithTests\ExampleTest
-  ✓ it example
+$snapshot = function ($name) {
+    $testsPath = dirname(__DIR__);
 
-  Tests:  1 passed
-EOF, $run('tests/Fixtures/DirectoryWithTests/ExampleTest.php'));
-});
+    return file_get_contents(implode(DIRECTORY_SEPARATOR, [
+        $testsPath,
+        '.snapshots',
+        "$name.txt",
+    ]));
+};
 
-test('allows to run a directory', function () use ($run) {
-    assertStringContainsString(<<<EOF
-   PASS  Tests\Fixtures\DirectoryWithTests\ExampleTest
-  ✓ it example
+test('allows to run a single test', function () use ($run, $snapshot) {
+    expect($run('tests/Fixtures/DirectoryWithTests/ExampleTest.php'))->toContain($snapshot('allows-to-run-a-single-test'));
+})->skip(PHP_OS_FAMILY === 'Windows');
 
-   PASS  Tests\Fixtures\ExampleTest
-  ✓ it example
+test('allows to run a directory', function () use ($run, $snapshot) {
+    expect($run('tests/Fixtures'))->toContain($snapshot('allows-to-run-a-directory'));
+})->skip(PHP_OS_FAMILY === 'Windows');
 
-  Tests:  2 passed
-EOF, $run('tests/Fixtures'));
-});
+it('has ascii chars', function () use ($run, $snapshot) {
+    expect($run('tests/Fixtures/DirectoryWithTests/ExampleTest.php', true))->toContain($snapshot('has-ascii-chars'));
+})->skip(PHP_OS_FAMILY === 'Windows');
+
+it('disable decorating printer when colors is set to never', function () use ($snapshot) {
+    $process = new Process([
+        'php',
+        './bin/pest',
+        '--colors=never',
+        'tests/Fixtures/DirectoryWithTests/ExampleTest.php',
+    ], dirname(__DIR__, 2));
+    $process->run();
+    $output = $process->getOutput();
+    expect($output)->toContain($snapshot('disable-decorating-printer'));
+})->skip(PHP_OS_FAMILY === 'Windows');