+++ tests/Visual/Success.php
@@ -9,7 +9,7 @@ test('visual snapshot of test suite on success', function () {
     ]);
 
     $output = function () use ($testsPath) {
-        $process = (new Symfony\Component\Process\Process(['./bin/pest'], dirname($testsPath), ['EXCLUDE' => 'integration', 'REBUILD_SNAPSHOTS' => false]));
+        $process = (new Symfony\Component\Process\Process(['php', 'bin/pest'], dirname($testsPath), ['EXCLUDE' => 'integration', 'REBUILD_SNAPSHOTS' => false]));
 
         $process->run();
 
@@ -17,11 +17,16 @@ test('visual snapshot of test suite on success', function () {
     };
 
     if (getenv('REBUILD_SNAPSHOTS')) {
-        file_put_contents($snapshot, $output());
+        // Strip time from end of snapshot
+        $outputContent = preg_replace('/Time\: \s+\d+\.\d+s\s+/m', '', $output());
+
+        file_put_contents($snapshot, $outputContent);
     } elseif (!getenv('EXCLUDE')) {
         $output = explode("\n", $output());
         array_pop($output);
         array_pop($output);
-        assertStringContainsString(implode("\n", $output), file_get_contents($snapshot));
+
+        expect(implode("\n", $output))->toContain(file_get_contents($snapshot));
     }
-})->skip(!getenv('REBUILD_SNAPSHOTS') && getenv('EXCLUDE'));
+})->skip(!getenv('REBUILD_SNAPSHOTS') && getenv('EXCLUDE'))
+    ->skip(PHP_OS_FAMILY === 'Windows');